Alpha_Creators_ads/backend/requirements_simple.txt
# Core FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database drivers (MongoDB only)
motor==3.3.2
pymongo==4.6.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pyjwt==2.8.0

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AI and ML (essential only)
openai==1.6.1

# HTTP and requests
httpx==0.25.2

# Cache (Redis - optional)
redis==5.0.1

# Email validation
email-validator==2.1.0

# Date/time utilities
python-dateutil==2.8.2